feat(compat): proptest L3D Pro parser + document Logitech yoke research
Research (linux-hardware.org, Linux kernel hid-ids.h, USB ID databases):
No Logitech-exclusive VID 0x046D PID confirmed for the Flight Yoke System.
The Logitech G-branded yoke is the same hardware as the Saitek Pro Flight
Yoke (VID 0x06A3, PID 0x0BAC), already covered by saitek/pro-flight-yoke.yaml
with vendor 'Saitek / Logitech'. Note added to that manifest.

Proptests (flight-hotas-logitech):
- arbitrary_bytes_all_axes_in_range: verifies all four axes stay within
  normalised bounds for arbitrary >=7 byte input (not just build_report-
  constructed reports), exercising bit-extraction arithmetic under
  adversarial patterns
- out_of_range_buttons_always_false: verifies button(0) and buttons 13-20
  always return false, and the high 4 bits of the 16-bit button word are
  always 0, regardless of raw report bytes

Also fix three pre-existing clippy::unnecessary_cast warnings in
parse_extreme_3d_pro (redundant 'as u8' on u8 expressions).

All 23 tests pass; cargo clippy -D warnings clean.

+++ crates/flight-hotas-logitech/src/extreme3dpro.rs
@@ -166,10 +166,10 @@ pub fn parse_extreme_3d_pro(data: &[u8]) -> Result<Extreme3DProInputState, Extre
     let y = ((data[1] as u16) >> 2) | ((data[2] as u16 & 0x0F) << 6);
 
     // Twist: bits 20-27
-    let twist = ((data[2] >> 4) | ((data[3] & 0x0F) << 4)) as u8;
+    let twist = (data[2] >> 4) | ((data[3] & 0x0F) << 4);
 
     // Throttle: bits 28-34 (7 bits)
-    let throttle_raw = ((data[3] >> 4) | ((data[4] & 0x07) << 4)) as u8;
+    let throttle_raw = (data[3] >> 4) | ((data[4] & 0x07) << 4);
 
     // Buttons: bits 35-46 (12 buttons)
     let btn_low = (data[4] >> 3) as u16; // 5 bits from byte 4
@@ -177,7 +177,7 @@ pub fn parse_extreme_3d_pro(data: &[u8]) -> Result<Extreme3DProInputState, Extre
     let buttons = btn_low | (btn_high << 5);
 
     // Hat: bits 47-50 (4 bits)
-    let hat_raw = ((data[5] >> 7) | ((data[6] & 0x07) << 1)) as u8;
+    let hat_raw = (data[5] >> 7) | ((data[6] & 0x07) << 1);
 
     Ok(Extreme3DProInputState {
         axes: Extreme3DProAxes {
